💄UI/UX: Enter Profile form when user selects apartment
#102

Not to let the user think twice about which one to choose on the home page.

Users can directly view a list of the apartments, whether they have profiles or not.

This would encourage users to look over the apartments at first.

If they are curious how much more money they need to buy the apartment.